+++ CHANGELOG.md
@@ -175,6 +175,7 @@ Full documentation for ROCprofiler-SDK is available at [rocm.docs.amd.com/projec
   - relative == logical_node_id
   - type-relative == logical_node_type_id
 - Added MI300 stochastic (hardware-based) PC sampling support in ROCProfiler-SDK and ROCProfV3
+- Python bindings for rocprofiler-sdk-roctx
 
 ### Changed